Fixed a bug that caused erroneous highlights in define statements
The goal of the define statement highlight rule is the emphasize the
place where a fully identified macro is created, changed, or destroyed.
It should only match when the first argument to the define-like macro
is [<identifier>].

To allow for undefine, popdef, and the single argument define, the
pattern for the comma after the first argument was optional. This was
an error, because in the case of `[<identifier>]<more>,' the parser
would simple match the end of the first quoted identifier and stop.

To fix this, the end of that pattern is now a non-optional forward
lookahead that must match a comma-or-close-paren character class.